+++ generador_primos.h
@@ -1,6 +1,6 @@
 #ifndef GENERAR_PRIMOS_H
 #define GENERAR_PRIMOS_H
 
-int* primos(int n);
+long long* primos(int n);
 
 #endif
